+++ Changelog.md
@@ -1,5 +1,7 @@
 # v1.3.0
 - Added "Hide Level Names" setting. All level names will be hidden when this setting is enabled; you will only know what levels have been selected when you enter them.
+  - (NOTE: Level names will still be displayed in votes regardless of if this setting is enabled.)
+- Added "Hide Player Times" setting. All times set by players will be hidden from view.
 - Fixed an issue where the player could enable cheats in a bingo game. You will now explode again and other players will get to laugh at you again.
 
 # v1.2.1